feat: integrate OAuth utilities and update plugin configurations for Google, Discord, and GitHub
