+++ packages/@studiocms/auth0/src/effect/auth0.ts
@@ -6,8 +6,8 @@ import { SDKCore } from 'studiocms:sdk';
 import { FetchHttpClient, HttpClient, HttpClientResponse } from '@effect/platform';
 import { Auth0, generateCodeVerifier, generateState } from 'arctic';
 import type { APIContext } from 'astro';
-import { AstroError } from 'astro/errors';
-import { Data, Effect, genLogger, pipe, Schema } from 'studiocms/effect';
+import { Effect, genLogger, pipe, Schema } from 'studiocms/effect';
+import { getCookie, getUrlParam, ValidateAuthCodeError } from 'studiocms/oAuthUtils';
 
 /**
  * Represents a user authenticated via Auth0.
@@ -27,35 +27,6 @@ export class Auth0User extends Schema.Class<Auth0User>('Auth0User')({
 }) {}
 
 /**
- * Retrieves the value of a specified query parameter from the given API context's URL.
- *
- * @param context - The API context containing the URL to extract the parameter from.
- * @param name - The name of the query parameter to retrieve.
- * @returns An Effect that resolves to the value of the query parameter, or throws an AstroError if parsing fails.
- */
-export const getUrlParam = ({ url }: APIContext, name: string) =>
-	Effect.try({
-		try: () => url.searchParams.get(name),
-		catch: () => new AstroError('Failed to parse URL from Astro context'),
-	});
-
-/**
- * Retrieves the value of a cookie from the provided API context using the specified key.
- *
- * Wraps the retrieval in an Effect, returning the cookie value if found, or `null` if not present.
- * Throws an `AstroError` if there is a failure during the cookie retrieval process.
- *
- * @param context - The API context containing the cookies object.
- * @param key - The name of the cookie to retrieve.
- * @returns An Effect that resolves to the cookie value as a string, or `null` if not found.
- */
-export const getCookie = ({ cookies }: APIContext, key: string) =>
-	Effect.try({
-		try: () => cookies.get(key)?.value ?? null,
-		catch: () => new AstroError('Failed to parse get Cookies from Astro context'),
-	});
-
-/**
  * Returns the normalized domain string for Auth0 authentication.
  *
  * This function performs the following transformations:
@@ -72,18 +43,6 @@ export const cleanDomain = (domain: string): string =>
 		(domain) => domain.replace(/(?:http|https):\/\//, ''),
 		(domain) => `https://${domain}`
 	);
-
-/**
- * Error class representing a failure during the validation of an authentication code.
- *
- * @extends Data.TaggedError
- * @property message - A descriptive error message explaining the cause of the failure.
- * @property provider - The authentication provider associated with the error (e.g., "auth0").
- */
-export class ValidateAuthCodeError extends Data.TaggedError('ValidateAuthCodeError')<{
-	message: string;
-	provider: string;
-}> {}
 
 /**
  * An object containing Auth0 configuration values retrieved from environment secrets.
