+++ packages/@studiocms/google/README.md
@@ -1,6 +1,6 @@
 # @studiocms/google Plugin
 
-Add Markdown support to StudioCMS
+This plugin integrates Google OAuth authentication into StudioCMS. It defines the necessary configuration, including the required environment variables, OAuth provider details, and the endpoint path for authentication.
 
 ## Usage
