+++ packages/@studiocms/google/src/index.ts
@@ -9,6 +9,23 @@
 import { createResolver } from 'astro-integration-kit';
 import { definePlugin, type StudioCMSPlugin } from 'studiocms/plugins';
 
+/**
+ * Creates and returns the StudioCMS Google Plugin.
+ *
+ * This plugin integrates Google OAuth authentication into StudioCMS.
+ * It defines the necessary configuration, including the required environment variables,
+ * OAuth provider details, and the endpoint path for authentication.
+ *
+ * @returns {StudioCMSPlugin} The configured StudioCMS Google Plugin instance.
+ *
+ * @remarks
+ * - Requires the following environment variables to be set:
+ *   - `CMS_GOOGLE_CLIENT_ID`
+ *   - `CMS_GOOGLE_CLIENT_SECRET`
+ *   - `CMS_GOOGLE_REDIRECT_URI`
+ * - Minimum supported StudioCMS version: `0.1.0-beta.22`
+ * - Registers the Google OAuth provider with a custom SVG logo.
+ */
 export function studiocmsGoogle(): StudioCMSPlugin {
 	// Resolve the path to the current file
 	const { resolve } = createResolver(import.meta.url);
